Fixed bug 614273: unable to delete global categories.

+++ category_handler.php
@@ -39,9 +39,16 @@ if ( ! $is_my_event )
 if ( empty ( $error ) &&
   ( $action == "Delete" || $action == translate ("Delete") ) ) {
   // delete this category
-  if ( ! dbi_query ( "DELETE FROM webcal_categories " .
-    "WHERE cat_id = $id AND cat_owner = '$login'" ) )
-    $error = translate ("Database error") . ": " . dbi_error();
+  if ( $is_admin ) {
+    if ( ! dbi_query ( "DELETE FROM webcal_categories " .
+      "WHERE cat_id = $id AND " .
+      "( cat_owner = '$login' OR cat_owner IS NULL )" ) )
+      $error = translate ("Database error") . ": " . dbi_error();
+  } else {
+    if ( ! dbi_query ( "DELETE FROM webcal_categories " .
+      "WHERE cat_id = $id AND cat_owner = '$login'" ) )
+      $error = translate ("Database error") . ": " . dbi_error();
+  }
       
   // Set any events in this category to NULL
   if ( ! dbi_query ( "UPDATE webcal_entry_user SET cal_category = NULL " .
